Refactored GDS event processing (#298)
* Took out GDS event processing from Pipeline package
* Harmonised domain objects between S3, GDS and Report
* Deployed `sqs_gds_event_processor` Lambda for GDS event queue
* Related #203
