+++ utils/tests/test_ica.py
@@ -4,6 +4,8 @@ from unittest import TestCase
 
 from mockito import unstub
 
+from utils.ica import GDSFilesEventType
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
@@ -25,6 +27,28 @@ class IcaUnitTests(TestCase):
         assert os.environ['ICA_BASE_URL'] == "http://localhost"
         assert os.environ['ICA_ACCESS_TOKEN'] == "mock"
         self.assertEqual(os.environ['ICA_BASE_URL'], "http://localhost")
+
+    def test_gds_files_event_type(self):
+        """
+        python -m unittest utils.tests.test_ica.IcaUnitTests.test_gds_files_event_type
+        """
+        event_type = GDSFilesEventType.from_value("deleted")
+        self.assertEqual(event_type, GDSFilesEventType.DELETED)
+
+        event_type = GDSFilesEventType.from_value("uploaded")
+        self.assertEqual(event_type, GDSFilesEventType.UPLOADED)
+
+        event_type = GDSFilesEventType.from_value("archived")
+        self.assertEqual(event_type, GDSFilesEventType.ARCHIVED)
+
+        event_type = GDSFilesEventType.from_value("unarchived")
+        self.assertEqual(event_type, GDSFilesEventType.UNARCHIVED)
+
+        try:
+            _ = GDSFilesEventType.from_value("unknown")
+        except Exception as e:
+            logger.exception(f"THIS ERROR EXCEPTION IS INTENTIONAL FOR TEST. NOT ACTUAL ERROR. \n{e}")
+        self.assertRaises(ValueError)
 
 
 class IcaIntegrationTests(TestCase):
